+++ controller/src/main/java/com/epam/esm/controller/CertificateController.java
@@ -1,8 +1,6 @@
 package com.epam.esm.controller;
 
-import com.epam.esm.domain.Tag;
 import com.epam.esm.dto.CertificateDto;
-import com.epam.esm.exception.CustomError;
 import com.epam.esm.exception.CustomExternalException;
 import com.epam.esm.exception.CustomNotFoundException;
 import com.epam.esm.exception.CustomNotValidArgumentException;
@@ -18,7 +16,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -29,7 +26,7 @@ import java.util.List;
  */
 @RestController
 @RequestMapping("/certificates")
-public class Controller{
+public class CertificateController {
 
     private static final String START_PATH = "/certificates/";
     private final CertificateService certificateService;
@@ -44,7 +41,7 @@ public class Controller{
      * @param certificateDtoValidator the certificate dto validator
      */
     @Autowired
-    public Controller(CertificateService certificateService, TagService tagService, CertificateDtoValidator certificateDtoValidator){
+    public CertificateController(CertificateService certificateService, TagService tagService, CertificateDtoValidator certificateDtoValidator){
         this.certificateService = certificateService;
         this.tagService = tagService;
         this.certificateDtoValidator = certificateDtoValidator;
@@ -91,44 +88,7 @@ public class Controller{
         return certificateService.showAll();
     }
 
-    /**
-     * Get tag.
-     *
-     * @param id the id
-     * @return the tag
-     */
-    @GetMapping(value = "/tag/{id}")
-    public Tag getTag(@PathVariable long id){
-        Tag tag = tagService.showById(id);
-        if(tag == null) {
-            throw  new CustomNotFoundException("Tag is not found by id = " + id);
-        }
-        return tag;
-    }
 
-    /**
-     * Get tag list.
-     *
-     * @return the list
-     */
-    @GetMapping(value = "/tag")
-    public List<Tag> getTag(){
-        return new ArrayList<>(tagService.showAll());
-    }
-
-    /**
-     * Delete tag.
-     *
-     * @param id the id
-     * @return the list
-     */
-    @DeleteMapping(value = "/tag/{id}")
-    public List<Tag> deleteTag(@PathVariable long id){
-        if(!tagService.delete(id)){
-            throw new CustomNotFoundException("Tag is not found by id = " + id);
-        }
-        return new ArrayList<>(tagService.showAll());
-    }
 
     /**
      * Delete certificate.
@@ -227,52 +187,5 @@ public class Controller{
         return certificateDto;
     }
 
-    /**
-     * Not found custom error. Http status 404.
-     *
-     * @param e the e
-     * @return the custom error
-     */
-    @ExceptionHandler(CustomNotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public @ResponseBody CustomError notFound(CustomNotFoundException e) {
-        return new CustomError(404, e.getMessage());
-    }
 
-    /**
-     * External error custom error. Http status 422.
-     *
-     * @param e the e
-     * @return the custom error
-     */
-    @ExceptionHandler(CustomNotValidArgumentException.class)
-    @ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
-    public @ResponseBody CustomError unprocessableError(CustomNotValidArgumentException e) {
-        return new CustomError(422, e.getMessage());
-    }
-
-    /**
-     * External error custom error. Http status 500.
-     *
-     * @param e the e
-     * @return the custom error
-     */
-    @ExceptionHandler(CustomExternalException.class)
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public @ResponseBody CustomError externalError(CustomExternalException e) {
-        return new CustomError(500, e.getMessage());
-    }
-
-    /**
-     * External error custom error. Http status 400. It is a global
-     * exception.
-     *
-     * @param e the e
-     * @return the custom error
-     */
-    @ExceptionHandler(Exception.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public @ResponseBody CustomError badRequestError(Exception e) {
-        return new CustomError(400, e.getMessage());
-    }
 }
